Use scheme and certificate
